+++ NEWS.md
@@ -4,10 +4,20 @@
 * `bartlettTests()` gains formula and lm methods
 * `colDevs()` examples extended
 
-## Version 1.7.1 (2024-05-03)
+## Version 1.7.2 (2024-11-25)
+
+* Fixed bug in `ellipse.axes()` re center/centre
+* `ellipse.axes()` gains a `type` argument to draw lines or arrows
+* `ellipse.axes()` gains a `extend` argument to extend/shrink the axes
+
+## Version 1.7.1 (2024-11-14)
+
+This is a modest update, adding two new datasets for simple examples and a PDF vignette
 
 * Extended the "datasets" vignette with further description of how this was done.
 * Added `dogfood` and `oral` datasets, both simple MANOVA examples.
+* Added HE Plots for Repeated Measures Designs vignette
+* Fixed nits from URLs vanished or moved
 
 ## Version 1.7.0 (2024-05-01)